+++ car-flex/MongoDB/publicacion_modificar_form.php
@@ -11,9 +11,11 @@
 
         if ((($_FILES["fotoModificada"]["type"] == "image/webp")) && ($_FILES["fotoModificada"]["size"])) {
             $archivo = $_FILES['fotoModificada']['name'];
-            $rutafotoModificada = 'img/fotosPerfil/' . time() . '_' . $NuevoUsuario . '_' . $archivo;
+            $rutafotoModificada = 'img/fotosPubli/' . time() . '_' . $NuevoUsuario . '_' . $archivo;
             $resultado = move_uploaded_file($_FILES["fotoModificada"]["tmp_name"], $rutafotoModificada);
+            if($_id_Publi!='img\fotosPubli\Foto_predeterminada.webp'){
             unlink($rutaModificada);
+            }
         } else {
             // Si no se sube una nueva foto, mantener la foto de perfil actual
             $rutafotoModificada = $_id_Publi; // Obtener la ruta de la foto de perfil actual desde la sesión
